Fix NextAuth configuration: centralize authOptions and fix circular imports
